+++ sdk/ml/azure-ai-ml/azure/ai/ml/constants/_job/job.py
@@ -107,7 +107,7 @@ class JobServiceTypeNames:
 
     REST_TO_ENTITY = {v: k for k, v in ENTITY_TO_REST.items()}
 
-    NAMES_ALLOWED_FOR_PUBLIC = [EntityNames.JUPYTER_LAB, EntityNames.SSH, EntityNames.TENSOR_BOARD, EntityNames.VS_CODE]
+    NAMES_ALLOWED_FOR_PUBLIC = [EntityNames.CUSTOM, EntityNames.TRACKING, EntityNames.STUDIO]
 
 
 class JobTierNames: