+++ sdk/identity/azure-identity/azure/identity/_credentials/client_secret.py
@@ -25,6 +25,15 @@ class ClientSecretCredential(ClientCredentialBase):
     :keyword List[str] additionally_allowed_tenants: Specifies tenants in addition to the specified "tenant_id"
         for which the credential may acquire tokens. Add the wildcard value "*" to allow the credential to
         acquire tokens for any tenant the application can access.
+
+    .. admonition:: Example:
+
+        .. literalinclude:: ../samples/credential_creation_code_snippets.py
+            :start-after: [START create_client_secret_credential]
+            :end-before: [END create_client_secret_credential]
+            :language: python
+            :dedent: 4
+            :caption: Create a ClientSecretCredential.
     """
 
     def __init__(