+++ sdk/ml/azure-ai-ml/azure/ai/ml/_schema/pipeline/settings.py
@@ -6,10 +6,10 @@
 
 from marshmallow import INCLUDE, Schema, fields, post_dump, post_load
 
-from azure.ai.ml._schema.core.fields import ArmStr
+from azure.ai.ml._schema.core.fields import ArmStr, StringTransformedEnum, UnionField
 from azure.ai.ml._schema.pipeline.pipeline_component import NodeNameStr
 from azure.ai.ml._utils.utils import is_private_preview_enabled
-from azure.ai.ml.constants._common import AzureMLResourceType
+from azure.ai.ml.constants._common import AzureMLResourceType, SERVERLESS_COMPUTE
 
 
 class PipelineJobSettingsSchema(Schema):
@@ -17,7 +17,12 @@ class PipelineJobSettingsSchema(Schema):
         unknown = INCLUDE
 
     default_datastore = ArmStr(azureml_type=AzureMLResourceType.DATASTORE)
-    default_compute = ArmStr(azureml_type=AzureMLResourceType.COMPUTE)
+    default_compute = UnionField(
+        [
+            StringTransformedEnum(allowed_values=[SERVERLESS_COMPUTE]),
+            ArmStr(azureml_type=AzureMLResourceType.COMPUTE),
+        ]
+    )
     continue_on_step_failure = fields.Bool()
     force_rerun = fields.Bool()